refactor(sessions): split runtime transcript projection
